requirements.txt
fastapi[all]
paho-mqtt